+++ CHANGELOG.md
@@ -79,6 +79,8 @@
 * Actually removed the exception I meant to remove in 0.31.0.
 * Changed `Attachment.type` to an enum instead of a string. This makes it easier to see all of the possible values.
 * Added option to allow attachment saving to be skipped when calling `Message.save`.
+* Added optional dependency `mimetype-magic` (installable using the `mime` extra) which helps to identify attachments that do not give a mime-type.
+* Added `type` property to all attachment types. `AttachmentBase` uses `AttachmentType.UNKNOWN`, `BrokenAttachment` uses `AttachmentType.BROKEN`, and `UnsupportedAttachment` uses `AttachmentType.UNSUPPORTED`.
 
 **v0.31.1**
 * Updated signed attachment mimetype property from `mime` to `mimetype` to match with the regular attachment property.